VLA-MoE-Manipulation/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
pyyaml>=6.0.1
tqdm>=4.66.0
einops>=0.7.0
numpy>=1.26.0
Pillow>=10.0.0
matplotlib>=3.8.0
